+++ test/storeTest.js
@@ -24,6 +24,7 @@ types.forEach(function (type) {
         expect(store.disconnect).to.be.a('function');
         expect(store.getNewId).to.be.a('function');
         expect(store.getEvents).to.be.a('function');
+        expect(store.getEventsSince).to.be.a('function');
         expect(store.getEventsByRevision).to.be.a('function');
         expect(store.getSnapshot).to.be.a('function');
         expect(store.addSnapshot).to.be.a('function');
@@ -886,6 +887,8 @@ types.forEach(function (type) {
 
           describe('having some events in the eventstore', function () {
 
+            var dateSince = new Date(Date.now() + 50);
+
             var stream1 = [{
               aggregateId: 'id',
               streamRevision: 0,
@@ -1113,6 +1116,463 @@ types.forEach(function (type) {
                   store.addEvents(stream10, callback);
                 }
               ], done);
+            });
+
+            describe('calling getEventsSince', function () {
+
+              describe('to get all events since a date', function () {
+
+                it('it should return the correct values', function (done) {
+
+                  var expectedEvts = allEvents.slice(4);
+
+                  store.getEventsSince(dateSince, 0, -1, function (err, evts) {
+                    expect(err).not.to.be.ok();
+                    expect(evts.length).to.eql(expectedEvts.length);
+
+                    var lastCommitStamp = 0;
+                    var lastCommitId = 0;
+                    var lastId = 0;
+                    _.each(evts, function (evt) {
+                      expect(evt.id).to.be.greaterThan(lastId);
+                      expect(evt.commitId >= lastCommitId).to.eql(true);
+                      expect(evt.commitStamp.getTime() >= lastCommitStamp).to.eql(true);
+                      lastId = evt.id;
+                      lastCommitId = evt.commitId;
+                      lastCommitStamp = evt.commitStamp.getTime();
+                    });
+
+                    done();
+                  });
+
+                });
+
+                describe('with a skip value', function () {
+
+                  it('it should return the correct values', function (done) {
+
+                    var expectedEvts = allEvents.slice(7);
+
+                    store.getEventsSince(dateSince, 3, -1, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(expectedEvts.length);
+
+                      var lastCommitStamp = 0;
+                      var lastCommitId = 0;
+                      var lastId = 0;
+                      _.each(evts, function (evt) {
+                        expect(evt.id).to.be.greaterThan(lastId);
+                        expect(evt.commitId >= lastCommitId).to.eql(true);
+                        expect(evt.commitStamp.getTime() >= lastCommitStamp).to.eql(true);
+                        lastId = evt.id;
+                        lastCommitId = evt.commitId;
+                        lastCommitStamp = evt.commitStamp.getTime();
+                      });
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+                describe('with a limit value', function () {
+
+                  it('it should return the correct values', function (done) {
+
+                    var expectedEvts = allEvents.slice(4, 9);
+
+                    store.getEventsSince(dateSince, 0, 5, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(expectedEvts.length);
+
+                      var lastCommitStamp = 0;
+                      var lastCommitId = 0;
+                      var lastId = 0;
+                      _.each(evts, function (evt) {
+                        expect(evt.id).to.be.greaterThan(lastId);
+                        expect(evt.commitId >= lastCommitId).to.eql(true);
+                        expect(evt.commitStamp.getTime() >= lastCommitStamp).to.eql(true);
+                        lastId = evt.id;
+                        lastCommitId = evt.commitId;
+                        lastCommitStamp = evt.commitStamp.getTime();
+                      });
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+                describe('with a skip and a limit value', function () {
+
+                  it('it should return the correct values', function (done) {
+
+                    var expectedEvts = allEvents.slice(7, 9);
+
+                    store.getEventsSince(dateSince, 4, 2, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(expectedEvts.length);
+
+                      var lastCommitStamp = 0;
+                      var lastCommitId = 0;
+                      var lastId = 0;
+                      _.each(evts, function (evt) {
+                        expect(evt.id).to.be.greaterThan(lastId);
+                        expect(evt.commitId >= lastCommitId).to.eql(true);
+                        expect(evt.commitStamp.getTime() >= lastCommitStamp).to.eql(true);
+                        lastId = evt.id;
+                        lastCommitId = evt.commitId;
+                        lastCommitStamp = evt.commitStamp.getTime();
+                      });
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+              });
+
+              describe('with an aggregateId being used only in one context and aggregate', function () {
+
+                it('it should return the correct events', function (done) {
+
+                  store.getEvents({ aggregateId: 'idWithAgg' }, 0, -1, function (err, evts) {
+                    expect(err).not.to.be.ok();
+                    expect(evts.length).to.eql(2);
+                    expect(evts[0].id).to.eql(stream2[0].id);
+                    expect(evts[0].aggregateId).to.eql(stream2[0].aggregateId);
+                    expect(evts[0].commitStamp.getTime()).to.eql(stream2[0].commitStamp.getTime());
+                    expect(evts[0].commitSequence).to.eql(stream2[0].commitSequence);
+                    expect(evts[0].streamRevision).to.eql(stream2[0].streamRevision);
+                    expect(evts[1].id).to.eql(stream2[1].id);
+                    expect(evts[1].aggregateId).to.eql(stream2[1].aggregateId);
+                    expect(evts[1].commitStamp.getTime()).to.eql(stream2[1].commitStamp.getTime());
+                    expect(evts[1].commitSequence).to.eql(stream2[1].commitSequence);
+                    expect(evts[1].streamRevision).to.eql(stream2[1].streamRevision);
+
+                    done();
+                  });
+
+                });
+
+                describe('and limit it with skip and limit', function () {
+
+                  it('it should return the correct events', function (done) {
+
+                    store.getEvents({ aggregateId: 'idWithAgg' }, 1, 2, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(1);
+                      expect(evts[0].aggregateId).to.eql(stream2[1].aggregateId);
+                      expect(evts[0].commitStamp.getTime()).to.eql(stream2[1].commitStamp.getTime());
+                      expect(evts[0].streamRevision).to.eql(stream2[1].streamRevision);
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+              });
+
+              describe('with an aggregateId being used in an other context or aggregate', function () {
+
+                it('it should return the correct events', function (done) {
+
+                  store.getEvents({ aggregateId: 'id' }, 0, -1, function (err, evts) {
+                    expect(err).not.to.be.ok();
+                    expect(evts.length).to.eql(5);
+                    expect(evts[0].aggregateId).to.eql(stream1[0].aggregateId);
+                    expect(evts[0].commitStamp.getTime()).to.eql(stream1[0].commitStamp.getTime());
+                    expect(evts[0].streamRevision).to.eql(stream1[0].streamRevision);
+                    expect(evts[1].aggregateId).to.eql(stream1[1].aggregateId);
+                    expect(evts[1].commitStamp.getTime()).to.eql(stream1[1].commitStamp.getTime());
+                    expect(evts[1].streamRevision).to.eql(stream1[1].streamRevision);
+                    expect(evts[2].aggregateId).to.eql(stream3[0].aggregateId);
+                    expect(evts[2].commitStamp.getTime()).to.eql(stream3[0].commitStamp.getTime());
+                    expect(evts[2].streamRevision).to.eql(stream3[0].streamRevision);
+                    expect(evts[3].aggregateId).to.eql(stream5[0].aggregateId);
+                    expect(evts[3].commitStamp.getTime()).to.eql(stream5[0].commitStamp.getTime());
+                    expect(evts[3].streamRevision).to.eql(stream5[0].streamRevision);
+                    expect(evts[4].aggregateId).to.eql(stream10[0].aggregateId);
+                    expect(evts[4].commitStamp.getTime()).to.eql(stream10[0].commitStamp.getTime());
+                    expect(evts[4].streamRevision).to.eql(stream10[0].streamRevision);
+
+                    done();
+                  });
+
+                });
+
+                describe('and limit it with revMin and revMax', function () {
+
+                  it('it should return the correct events', function (done) {
+
+                    store.getEvents({ aggregateId: 'id' }, 1, 2, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(2);
+                      expect(evts[0].aggregateId).to.eql(stream1[1].aggregateId);
+                      expect(evts[0].commitStamp.getTime()).to.eql(stream1[1].commitStamp.getTime());
+                      expect(evts[0].streamRevision).to.eql(stream1[1].streamRevision);
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+              });
+
+              describe('without an aggregateId but with an aggregate', function () {
+
+                it('it should return the correct events', function (done) {
+
+                  store.getEvents({ aggregate: 'myAggrrr2' }, 0, -1, function (err, evts) {
+                    expect(err).not.to.be.ok();
+                    expect(evts.length).to.eql(3);
+                    expect(evts[0].aggregateId).to.eql(stream7[0].aggregateId);
+                    expect(evts[0].commitStamp.getTime()).to.eql(stream7[0].commitStamp.getTime());
+                    expect(evts[0].streamRevision).to.eql(stream7[0].streamRevision);
+                    expect(evts[1].aggregateId).to.eql(stream7[1].aggregateId);
+                    expect(evts[1].commitStamp.getTime()).to.eql(stream7[1].commitStamp.getTime());
+                    expect(evts[1].streamRevision).to.eql(stream7[1].streamRevision);
+                    expect(evts[2].aggregateId).to.eql(stream9[0].aggregateId);
+                    expect(evts[2].commitStamp.getTime()).to.eql(stream9[0].commitStamp.getTime());
+                    expect(evts[2].streamRevision).to.eql(stream9[0].streamRevision);
+
+                    done();
+                  });
+
+                });
+
+                describe('and limit it with skip and limit', function () {
+
+                  it('it should return the correct events', function (done) {
+
+                    store.getEvents({ aggregate: 'myAggrrr2' }, 1, 2, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(2);
+                      expect(evts[0].aggregateId).to.eql(stream7[1].aggregateId);
+                      expect(evts[0].commitStamp.getTime()).to.eql(stream7[1].commitStamp.getTime());
+                      expect(evts[0].streamRevision).to.eql(stream7[1].streamRevision);
+                      expect(evts[1].aggregateId).to.eql(stream9[0].aggregateId);
+                      expect(evts[1].commitStamp.getTime()).to.eql(stream9[0].commitStamp.getTime());
+                      expect(evts[1].streamRevision).to.eql(stream9[0].streamRevision);
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+              });
+
+              describe('with an aggregateId and with an aggregate', function () {
+
+                it('it should return the correct events', function (done) {
+
+                  store.getEvents({ aggregate: 'myAggrrr2', aggregateId: 'idWithAggrAndCont' }, 0, -1, function (err, evts) {
+                    expect(err).not.to.be.ok();
+                    expect(evts.length).to.eql(1);
+                    expect(evts[0].aggregateId).to.eql(stream9[0].aggregateId);
+                    expect(evts[0].commitStamp.getTime()).to.eql(stream9[0].commitStamp.getTime());
+                    expect(evts[0].streamRevision).to.eql(stream9[0].streamRevision);
+
+                    done();
+                  });
+
+                });
+
+                describe('and limit it with skip and limit', function () {
+
+                  it('it should return the correct events', function (done) {
+
+                    store.getEvents({ aggregate: 'myAggrrr2', aggregateId: 'idWithAggrAndCont' }, 1, 2, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(0);
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+              });
+
+              describe('with an aggregateId and without an aggregate but with a context', function () {
+
+                it('it should return the correct events', function (done) {
+
+                  store.getEvents({ aggregateId: 'idWithAggrAndCont', context: 'myConttttt' }, 0, -1, function (err, evts) {
+                    expect(err).not.to.be.ok();
+                    expect(evts.length).to.eql(3);
+                    expect(evts[0].aggregateId).to.eql(stream6[0].aggregateId);
+                    expect(evts[0].commitStamp.getTime()).to.eql(stream6[0].commitStamp.getTime());
+                    expect(evts[0].streamRevision).to.eql(stream6[0].streamRevision);
+                    expect(evts[1].aggregateId).to.eql(stream6[1].aggregateId);
+                    expect(evts[1].commitStamp.getTime()).to.eql(stream6[1].commitStamp.getTime());
+                    expect(evts[1].streamRevision).to.eql(stream6[1].streamRevision);
+                    expect(evts[2].aggregateId).to.eql(stream9[0].aggregateId);
+                    expect(evts[2].commitStamp.getTime()).to.eql(stream9[0].commitStamp.getTime());
+                    expect(evts[2].streamRevision).to.eql(stream9[0].streamRevision);
+
+                    done();
+                  });
+
+                });
+
+                describe('and limit it with skip and limit', function () {
+
+                  it('it should return the correct events', function (done) {
+
+                    store.getEvents({ aggregateId: 'idWithAggrAndCont', context: 'myConttttt' }, 1, 2, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(2);
+                      expect(evts[0].aggregateId).to.eql(stream6[1].aggregateId);
+                      expect(evts[0].commitStamp.getTime()).to.eql(stream6[1].commitStamp.getTime());
+                      expect(evts[0].streamRevision).to.eql(stream6[1].streamRevision);
+                      expect(evts[1].aggregateId).to.eql(stream9[0].aggregateId);
+                      expect(evts[1].commitStamp.getTime()).to.eql(stream9[0].commitStamp.getTime());
+                      expect(evts[1].streamRevision).to.eql(stream9[0].streamRevision);
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+              });
+
+              describe('with an aggregateId and with an aggregate and with a context', function () {
+
+                it('it should return the correct events', function (done) {
+
+                  store.getEvents({ aggregateId: 'id', aggregate: 'wowAgg', context: 'wowCont' }, 0, -1, function (err, evts) {
+                    expect(err).not.to.be.ok();
+                    expect(evts.length).to.eql(1);
+                    expect(evts[0].aggregateId).to.eql(stream10[0].aggregateId);
+                    expect(evts[0].commitStamp.getTime()).to.eql(stream10[0].commitStamp.getTime());
+                    expect(evts[0].streamRevision).to.eql(stream10[0].streamRevision);
+
+                    done();
+                  });
+
+                });
+
+                describe('and limit it with skip and limit', function () {
+
+                  it('it should return the correct events', function (done) {
+
+                    store.getEvents({ aggregateId: 'id', aggregate: 'wowAgg', context: 'wowCont' }, 1, 2, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(0);
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+              });
+
+              describe('without an aggregateId and without an aggregate but with a context', function () {
+
+                it('it should return the correct events', function (done) {
+
+                  store.getEvents({ context: 'myCont' }, 0, -1, function (err, evts) {
+                    expect(err).not.to.be.ok();
+                    expect(evts.length).to.eql(3);
+                    expect(evts[0].aggregateId).to.eql(stream4[0].aggregateId);
+                    expect(evts[0].commitStamp.getTime()).to.eql(stream4[0].commitStamp.getTime());
+                    expect(evts[0].streamRevision).to.eql(stream4[0].streamRevision);
+                    expect(evts[1].aggregateId).to.eql(stream4[1].aggregateId);
+                    expect(evts[1].commitStamp.getTime()).to.eql(stream4[1].commitStamp.getTime());
+                    expect(evts[1].streamRevision).to.eql(stream4[1].streamRevision);
+                    expect(evts[2].aggregateId).to.eql(stream5[0].aggregateId);
+                    expect(evts[2].commitStamp.getTime()).to.eql(stream5[0].commitStamp.getTime());
+                    expect(evts[2].streamRevision).to.eql(stream5[0].streamRevision);
+
+                    done();
+                  });
+
+                });
+
+                describe('and limit it with skip and limit', function () {
+
+                  it('it should return the correct events', function (done) {
+
+                    store.getEvents({ context: 'myCont' }, 1, 2, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(2);
+                      expect(evts[0].aggregateId).to.eql(stream4[1].aggregateId);
+                      expect(evts[0].commitStamp.getTime()).to.eql(stream4[1].commitStamp.getTime());
+                      expect(evts[0].streamRevision).to.eql(stream4[1].streamRevision);
+                      expect(evts[1].aggregateId).to.eql(stream5[0].aggregateId);
+                      expect(evts[1].commitStamp.getTime()).to.eql(stream5[0].commitStamp.getTime());
+                      expect(evts[1].streamRevision).to.eql(stream5[0].streamRevision);
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+              });
+
+              describe('without an aggregateId but with an aggregate and with a context', function () {
+
+                it('it should return the correct events', function (done) {
+
+                  store.getEvents({ context: 'myConttttt', aggregate: 'myAggrrr' }, 0, -1, function (err, evts) {
+                    expect(err).not.to.be.ok();
+                    expect(evts.length).to.eql(3);
+                    expect(evts[0].aggregateId).to.eql(stream6[0].aggregateId);
+                    expect(evts[0].commitStamp.getTime()).to.eql(stream6[0].commitStamp.getTime());
+                    expect(evts[0].streamRevision).to.eql(stream6[0].streamRevision);
+                    expect(evts[1].aggregateId).to.eql(stream6[1].aggregateId);
+                    expect(evts[1].commitStamp.getTime()).to.eql(stream6[1].commitStamp.getTime());
+                    expect(evts[1].streamRevision).to.eql(stream6[1].streamRevision);
+                    expect(evts[2].aggregateId).to.eql(stream8[0].aggregateId);
+                    expect(evts[2].commitStamp.getTime()).to.eql(stream8[0].commitStamp.getTime());
+                    expect(evts[2].streamRevision).to.eql(stream8[0].streamRevision);
+
+                    done();
+                  });
+
+                });
+
+                describe('and limit it with skip and limit', function () {
+
+                  it('it should return the correct events', function (done) {
+
+                    store.getEvents({ context: 'myConttttt', aggregate: 'myAggrrr' }, 1, 2, function (err, evts) {
+                      expect(err).not.to.be.ok();
+                      expect(evts.length).to.eql(2);
+                      expect(evts[0].aggregateId).to.eql(stream6[1].aggregateId);
+                      expect(evts[0].commitStamp.getTime()).to.eql(stream6[1].commitStamp.getTime());
+                      expect(evts[0].streamRevision).to.eql(stream6[1].streamRevision);
+                      expect(evts[1].aggregateId).to.eql(stream8[0].aggregateId);
+                      expect(evts[1].commitStamp.getTime()).to.eql(stream8[0].commitStamp.getTime());
+                      expect(evts[1].streamRevision).to.eql(stream8[0].streamRevision);
+
+                      done();
+                    });
+
+                  });
+
+                });
+
+              });
+
             });
 
             describe('calling getEvents', function () {